+++ Backend/Database/Repositories/UserRepository.cs
@@ -14,7 +14,7 @@ public class UserRepository : IUserRepository
     public async Task<IEnumerable<User>> GetAllAsync() =>
         await _context.Users.ToListAsync();
 
-    public async Task<User?> GetByIdAsync(long id) =>
+    public async Task<User?> GetByIdAsync(int id) =>
         await _context.Users.FindAsync(id);
 
     public async void DeleteAsync(User entity)
